[MODCON-94] - Add required permissions during setup only for ECS mode
