+++ src/test/java/org/folio/consortia/service/PermissionServiceTest.java
@@ -1,20 +1,13 @@
 package org.folio.consortia.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.folio.consortia.utils.EntityUtils.RANDOM_USER_ID;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import java.util.List;
-import java.util.UUID;
-
 import org.folio.consortia.client.PermissionsClient;
-import org.folio.consortia.domain.dto.PermissionUser;
-import org.folio.consortia.service.impl.PermissionUserServiceImpl;
-import org.junit.jupiter.api.Assertions;
+import org.folio.consortia.domain.dto.Permission;
+import org.folio.consortia.service.impl.PermissionServiceImpl;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -24,49 +17,20 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 @EnableAutoConfiguration(exclude = BatchAutoConfiguration.class)
 class PermissionServiceTest {
-  private static final String PERMISSIONS_FILE_PATH = "permissions/test-user-permissions.csv";
-  private static final String EMPTY_PERMISSIONS_FILE_PATH = "permissions/test-user--empty-permissions.csv";
   @InjectMocks
-  PermissionUserServiceImpl permissionUserService;
+  PermissionServiceImpl permissionService;
   @Mock
   PermissionsClient permissionsClient;
 
   @Test
-  void shouldThrowErrorForEmptyPermissionFileWhileAdding() {
-    PermissionUser permissionUser = PermissionUser.of(UUID.randomUUID().toString(), RANDOM_USER_ID, List.of());
+  void shouldCreatePermission() {
+    String permissionName = "consortia.consortia-configuration.item.post";
+    Permission permission = new Permission(permissionName);
 
-    Assertions.assertThrows(java.lang.IllegalStateException.class, () -> permissionUserService.addPermissions(permissionUser, EMPTY_PERMISSIONS_FILE_PATH));
+    doNothing().when(permissionsClient).createPermission(permission);
+
+    permissionService.createPermission(permissionName);
+
+    verify(permissionService, times(1)).createPermission(permissionName);
   }
-
-  @Test
-  void shouldThrowErrorForEmptyPermissionFileWhileCreating() {
-    Assertions.assertThrows(java.lang.IllegalStateException.class, () -> permissionUserService.createWithPermissionsFromFile(UUID.randomUUID().toString(), EMPTY_PERMISSIONS_FILE_PATH));
-  }
-
-  @Test
-  void shouldAddPermissionsToPermissionUser() {
-    PermissionUser permissionUser = PermissionUser.of(UUID.randomUUID().toString(), RANDOM_USER_ID, List.of());
-
-    doNothing().when(permissionsClient).addPermission(any(),any());
-    Assertions.assertDoesNotThrow(() -> permissionUserService.addPermissions(permissionUser, PERMISSIONS_FILE_PATH));
-  }
-
-  @Test
-  void shouldCreateEmptyPermission() {
-    String userId = RANDOM_USER_ID;
-    ArgumentCaptor<PermissionUser> captor = ArgumentCaptor.forClass(PermissionUser.class);
-    permissionUserService.createWithEmptyPermissions(userId);
-    verify(permissionsClient).create(captor.capture());
-    PermissionUser created = captor.getValue();
-    assertThat(created.getUserId()).isEqualTo(userId);
-    assertThat(created.getPermissions()).isEmpty();
-  }
-
-  @Test
-  void shouldDeletePermissionUser() {
-    String permissionUserId = UUID.randomUUID().toString();
-    permissionUserService.deletePermissionUser(permissionUserId);
-    verify(permissionsClient).deletePermissionUser(permissionUserId);
-  }
-
 }
